Separating `ClearActiveFiltersLink` component
